Prometheus operator yaml uses new crd apiVersion
The prometheus operator publishes yaml updates frequently and we would
like to be up to date in out documentation. This means we install the
same bundle.yaml file that they keep up to dote.

The bundle.yaml file diverged and changed the apiVersion of the CRD to
apiextensions.k8s.io/v1 which does not exist in 1.14 and below. Google
defaults to using 1.14 so we need to have a workaround in our
documentation.

Signed-off-by: Noah Krause <krausenoah@gmail.com>